5289 temperature.include: Allow floating point numbers in upper and lower levels in 'Temperature' ruleset

+++ web/plugins/wato/check_parameters.py
@@ -12176,16 +12176,16 @@ register_check_parameters(
                   Tuple(
                       title = _("Upper Temperature Levels"),
                       elements = [
-                          Integer(title = _("Warning at"), unit = u"°C", default_value = 26),
-                          Integer(title = _("Critical at"), unit = u"°C", default_value = 30),
+                          Float(title = _("Warning at"), unit = u"°C", default_value = 26),
+                          Float(title = _("Critical at"), unit = u"°C", default_value = 30),
                       ]
                 )),
                 ( "levels_lower",
                   Tuple(
                       title = _("Lower Temperature Levels"),
                       elements = [
-                          Integer(title = _("Warning below"), unit = u"°C", default_value = 0),
-                          Integer(title = _("Critical below"), unit = u"°C", default_value = -10),
+                          Float(title = _("Warning below"), unit = u"°C", default_value = 0),
+                          Float(title = _("Critical below"), unit = u"°C", default_value = -10),
                       ]
                 )),
                 ( "output_unit",
